+++ examples/decimal_number.go
@@ -37,7 +37,7 @@ func main() {
 	}
 
 	examples := []struct {
-		number  int8
+		number  int32
 		decimal uint8
 	}{
 		{42, 0},
